Refactor the framework module - Move Component and Condition in respective module - Move meta-components into control_flow and logic modules
