+++ src/components/generation/mod.rs
@@ -1,104 +1,20 @@
 //! Generation methods
 
-use crate::{
-    framework::{
-        components::{AnyComponent, Component},
-        Individual,
-    },
-    problems::{LimitedVectorProblem, Problem, VectorProblem},
-    state::State,
-};
 use itertools::Itertools;
 use rand::distributions::uniform::SampleUniform;
 use serde::{Deserialize, Serialize};
 
+use crate::{
+    components::Component,
+    problems::{LimitedVectorProblem, Problem, VectorProblem},
+    state::State,
+};
+
 pub mod mutation;
+pub use mutation::{Generation, Generator};
+
 pub mod recombination;
 pub mod swarm;
-
-/// Specialized component trait to generate a new population from the current one.
-///
-/// This trait is especially useful for components that modify solutions independently.
-/// For combining multiple solutions, see [Recombination].
-///
-/// # Implementing [Component]
-///
-/// Types implementing this trait can implement [Component] by wrapping the type in a [Generator].
-pub trait Generation<P: Problem> {
-    fn generate_population(
-        &self,
-        population: &mut Vec<P::Encoding>,
-        problem: &P,
-        state: &mut State<P>,
-    );
-}
-
-#[derive(serde::Serialize, Clone)]
-pub struct Generator<T: Clone>(pub T);
-
-impl<T, P> Component<P> for Generator<T>
-where
-    P: Problem,
-    T: AnyComponent + Generation<P> + Serialize + Clone,
-{
-    fn execute(&self, problem: &P, state: &mut State<P>) {
-        let population = state.populations_mut().pop();
-        let mut population = population
-            .into_iter()
-            .map(Individual::into_solution)
-            .collect();
-        self.0.generate_population(&mut population, problem, state);
-        let population = population
-            .into_iter()
-            .map(Individual::<P>::new_unevaluated)
-            .collect();
-        state.populations_mut().push(population);
-    }
-}
-
-/// Specialized component trait to generate a new population from the current one.
-///
-/// This trait is especially useful for components that combine multiple solutions.
-/// For modifying solutions independently, see [Generation].
-///
-/// # Implementing [Component]
-///
-/// Types implementing this trait can implement [Component] by wrapping the type in a [Recombinator].
-pub trait Recombination<P: Problem> {
-    fn recombine_solutions(
-        &self,
-        parents: Vec<P::Encoding>,
-        offspring: &mut Vec<P::Encoding>,
-        problem: &P,
-        state: &mut State<P>,
-    );
-}
-
-#[derive(serde::Serialize, Clone)]
-pub struct Recombinator<T: Clone>(pub T);
-
-impl<T, P, D> Component<P> for Recombinator<T>
-where
-    P: Problem<Encoding = Vec<D>>,
-    T: AnyComponent + Recombination<P> + Serialize + Clone,
-    D: Clone + PartialEq + 'static,
-{
-    fn execute(&self, problem: &P, state: &mut State<P>) {
-        let population = state.populations_mut().pop();
-        let population = population
-            .into_iter()
-            .map(Individual::into_solution)
-            .collect();
-        let mut offspring = Vec::new();
-        self.0
-            .recombine_solutions(population, &mut offspring, problem, state);
-        let offspring = offspring
-            .into_iter()
-            .map(Individual::<P>::new_unevaluated)
-            .collect();
-        state.populations_mut().push(offspring);
-    }
-}
 
 // Random Operators without state
