+++ src/components/replacement.rs
@@ -1,12 +1,14 @@
 //! Replacement methods
 
+use rand::seq::SliceRandom;
+use serde::{Deserialize, Serialize};
+
 use crate::{
-    framework::{components::*, Individual},
+    components::Component,
+    framework::{AnyComponent, Individual},
     problems::{Problem, SingleObjectiveProblem},
     state::State,
 };
-use rand::seq::SliceRandom;
-use serde::{Deserialize, Serialize};
 
 /// Specialized component trait to replace the population with the child population,
 /// typically generated by a [Selection](crate::components::selection::Selection) component.
@@ -44,11 +46,13 @@ where
 /// Discards all individuals in the child population, keeping the parents unchanged.
 #[derive(Serialize, Deserialize, Clone)]
 pub struct Noop;
+
 impl Noop {
     pub fn new<P: Problem>() -> Box<dyn Component<P>> {
         Box::new(Replacer(Self))
     }
 }
+
 impl<P: Problem> Replacement<P> for Noop {
     fn replace_population(
         &self,
@@ -65,6 +69,7 @@ pub struct MuPlusLambda {
     /// Limits the population growth.
     pub max_population_size: u32,
 }
+
 impl MuPlusLambda {
     pub fn new<P: SingleObjectiveProblem>(max_population_size: u32) -> Box<dyn Component<P>> {
         Box::new(Replacer(Self {
@@ -72,6 +77,7 @@ impl MuPlusLambda {
         }))
     }
 }
+
 impl<P: SingleObjectiveProblem> Replacement<P> for MuPlusLambda {
     fn replace_population(
         &self,
@@ -87,7 +93,6 @@ impl<P: SingleObjectiveProblem> Replacement<P> for MuPlusLambda {
 
 #[cfg(test)]
 mod mupluslambda {
-    use crate::state::State;
     use crate::testing::*;
 
     use super::*;
@@ -135,8 +140,9 @@ impl<P: Problem> Replacement<P> for Generational {
 
 #[cfg(test)]
 mod generational {
-    use super::*;
     use crate::testing::*;
+
+    use super::*;
 
     #[test]
     fn keeps_all_children() {
@@ -181,9 +187,10 @@ impl<P: Problem> Replacement<P> for RandomReplacement {
 
 #[cfg(test)]
 mod random_replacement {
-    use super::*;
-    use crate::framework::Random;
+    use crate::state::Random;
     use crate::testing::*;
+
+    use super::*;
 
     #[test]
     fn keeps_right_amount_of_children() {
@@ -226,9 +233,10 @@ impl<P: SingleObjectiveProblem> Replacement<P> for IndividualPlus {
 }
 #[cfg(test)]
 mod greedy_index {
-    use super::*;
-    use crate::framework::Random;
+    use crate::state::Random;
     use crate::testing::*;
+
+    use super::*;
 
     #[test]
     fn keeps_right_amount_of_children() {
