+++ src/conditions/logic.rs
@@ -1,33 +1,20 @@
 use derivative::Derivative;
 use serde::Serialize;
 
-/// The Condition trait and combinators.
-use crate::{framework::components::AnyComponent, problems::Problem, state::State};
-
-/// A condition for loops or branches.
-///
-/// Similar to [Component](crate::framework::components::Component),
-/// but `evaluate` replaces `execute` and returns a `bool`.
-///
-/// These can be combined using binary AND and OR (`|` and `&`).
-pub trait Condition<P: Problem>: AnyComponent {
-    #[allow(unused_variables)]
-    fn initialize(&self, problem: &P, state: &mut State<P>) {}
-    fn evaluate(&self, problem: &P, state: &mut State<P>) -> bool;
-}
-erased_serde::serialize_trait_object!(<P: Problem> Condition<P>);
-dyn_clone::clone_trait_object!(<P: Problem> Condition<P>);
+use crate::{conditions::Condition, problems::Problem, state::State};
 
 /// Multiple conditions must be true.
 #[derive(Serialize, Derivative)]
 #[serde(bound = "")]
 #[derivative(Clone(bound = ""))]
 pub struct And<P: Problem>(Vec<Box<dyn Condition<P>>>);
+
 impl<P: Problem> And<P> {
     pub fn new(conditions: Vec<Box<dyn Condition<P>>>) -> Box<dyn Condition<P>> {
         Box::new(Self(conditions))
     }
 }
+
 impl<P: Problem> Condition<P> for And<P> {
     fn initialize(&self, problem: &P, state: &mut State<P>) {
         for condition in self.0.iter() {
@@ -41,6 +28,7 @@ impl<P: Problem> Condition<P> for And<P> {
             .all(|condition| condition.evaluate(problem, state))
     }
 }
+
 impl<P: Problem> std::ops::BitAnd for Box<dyn Condition<P>> {
     type Output = Box<dyn Condition<P>>;
 
@@ -54,11 +42,13 @@ impl<P: Problem> std::ops::BitAnd for Box<dyn Condition<P>> {
 #[serde(bound = "")]
 #[derivative(Clone(bound = ""))]
 pub struct Or<P: Problem>(Vec<Box<dyn Condition<P>>>);
+
 impl<P: Problem> Or<P> {
     pub fn new(conditions: Vec<Box<dyn Condition<P>>>) -> Box<dyn Condition<P>> {
         Box::new(Self(conditions))
     }
 }
+
 impl<P: Problem> Condition<P> for Or<P> {
     fn initialize(&self, problem: &P, state: &mut State<P>) {
         for condition in self.0.iter() {
@@ -72,6 +62,7 @@ impl<P: Problem> Condition<P> for Or<P> {
             .any(|condition| condition.evaluate(problem, state))
     }
 }
+
 impl<P: Problem> std::ops::BitOr for Box<dyn Condition<P>> {
     type Output = Box<dyn Condition<P>>;
 
@@ -85,11 +76,13 @@ impl<P: Problem> std::ops::BitOr for Box<dyn Condition<P>> {
 #[serde(bound = "")]
 #[derivative(Clone(bound = ""))]
 pub struct Not<P: Problem>(Box<dyn Condition<P>>);
+
 impl<P: Problem> Not<P> {
     pub fn new(condition: Box<dyn Condition<P>>) -> Box<dyn Condition<P>> {
         Box::new(Self(condition))
     }
 }
+
 impl<P: Problem> Condition<P> for Not<P> {
     fn initialize(&self, problem: &P, state: &mut State<P>) {
         self.0.initialize(problem, state);
@@ -99,6 +92,7 @@ impl<P: Problem> Condition<P> for Not<P> {
         !self.0.evaluate(problem, state)
     }
 }
+
 impl<P: Problem> std::ops::Not for Box<dyn Condition<P>> {
     type Output = Box<dyn Condition<P>>;
