+++ src/heuristics/ga.rs
@@ -1,9 +1,9 @@
 //! Genetic Algorithm
 
-use crate::prelude::branching;
 use crate::{
     components::*,
-    framework::{components::Component, conditions::Condition, Configuration},
+    conditions::*,
+    framework::Configuration,
     problems::{LimitedVectorProblem, SingleObjectiveProblem, VectorProblem},
 };
 
@@ -148,7 +148,7 @@ pub fn ga<P: SingleObjectiveProblem>(
                 .do_(constraints)
                 .evaluate()
                 .update_best_individual()
-                .do_optional_(archive)
+                .do_if_some_(archive)
                 .do_(replacement)
                 .do_(logger)
         })
