+++ src/state/pso.rs
@@ -1,10 +1,12 @@
+use better_any::Tid;
+use rand::Rng;
+
 use crate::{
-    framework::{components::*, Individual},
+    components::Component,
+    framework::Individual,
     problems::{LimitedVectorProblem, Problem, SingleObjectiveProblem},
     state::{CustomState, State},
 };
-use better_any::Tid;
-use rand::Rng;
 
 /// State required for PSO.
 ///
